Fixing Bugs, Redirect Page for Convenient UX

+++ model.php
@@ -27,24 +27,37 @@
 
 	function makeQuestionView($id) {
 		$question = getQuestionbyId($id);
+		if ( !empty($question) ) {
+			echo("<div class='title'> " . $question['topic'] . " </div>");
 
-		echo("<div class='title'> " . $question['topic'] . " </div>" . 
-			 " <div class='qview'> " .
-			 " 		<div class='votes'>" . 
-			 "			<a class='up_vote' onclick=\"voting(" . $question['id'] . ", 'up', 'question')\"></a>" .
-			 "			<div id='vote_count_q_" . $question['id'] . "' class='vote_count'>" . $question['vote'] . "</div>" .
-			 "			<a class='down_vote' onclick=\"voting(" . $question['id'] . ", 'down', 'question')\"></a>" .
-			 "		</div>" .
-			 " 		<div class='data'> " .
-			 " 			<div class='content'>" . $question['content'] . "</div>" .
-			 " 			<div class='control'>asked by " . $question['name'] . " at " . $question['date_created'] . " | <a class='edit' href='/edit/question/" . $question['id'] . "'> edit </a> | <a class='delete' onclick='deleteQuestion(" . $question['id'] . ")'>delete</a> </div>" .
-			 " 		</div> " .
-			 " </div> ");		
+			if ( $question['vote'] >= 0 ) {
+				echo(" <div class='qview'> ");
+			} else {
+				echo(" <div class='qview negative'> ");
+			}
+
+			echo(" 		<div class='votes'>" . 
+				 "			<a class='up_vote' onclick=\"voting(" . $question['id'] . ", 'up', 'question')\"></a>" .
+				 "			<div id='vote_count_q_" . $question['id'] . "' class='vote_count'>" . $question['vote'] . "</div>" .
+				 "			<a class='down_vote' onclick=\"voting(" . $question['id'] . ", 'down', 'question')\"></a>" .
+				 "		</div>" .
+				 " 		<div class='data'> " .
+				 " 			<div class='content'>" . $question['content'] . "</div>" .
+				 " 			<div class='control'>asked by " . $question['name'] . " at " . $question['date_created'] . " | <a class='edit' href='/edit/question/" . $question['id'] . "'> edit </a> | <a class='delete' onclick='deleteQuestion(" . $question['id'] . ")'>delete</a> </div>" .
+				 " 		</div> " .
+				 " </div> ");	
+		} else {
+			redirectTo("/home");
+		}	
 	}
 
 	function makeAnswerView($answer) {
-		echo(" <div class='aview'> " .
-			 " 		<div class='votes'>" . 
+		if ( $answer['vote'] >= 0 ) {
+			echo(" <div class='aview'> ");
+		} else {
+			echo(" <div class='aview negative'> ");
+		}
+		echo(" 		<div class='votes'>" . 
 			 "			<a class='up_vote' onclick=\"voting(" . $answer['id'] . ", 'up', 'answer')\"></a>" .
 			 "			<div id='vote_count_a_" . $answer['id'] . "' class='vote_count'>" . $answer['vote'] . "</div>" .
 			 "			<a class='down_vote' onclick=\"voting(" . $answer['id'] . ", 'down', 'answer')\"></a>" .
